Move process_mimikatz to core
